Update avscan.sh
Haved hashed out the line : echo "Content-Type: text/html" : in avsacn.sh as there is an issue with FreeNAS v11.3 and sendmail.  With this line enabled the email will not be sent, but if you has (#) it out they are.  For now have just hashed it out, so the email works.